# swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "version": "1.0.0",
    "description": "API for E-Commerce application with 404 handler for undefined routes",
    "title": "E-commerce API"
  },
  "servers": [
    {
      "url": "http://localhost:3200",
      "description": "Development server"
    }
  ],
  "components": {
    "securitySchemes": {
      "JWT": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "JWT token for authentication"
      }
    },
    "schemas": {
      "Product": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "example": 1
          },
          "name": {
            "type": "string",
            "example": "Wireless Bluetooth Headphones"
          },
          "desc": {
            "type": "string",
            "example": "High-quality noise-cancelling headphones with 30-hour battery life"
          },
          "imageUrl": {
            "type": "string",
            "example": "https://images.unsplash.com/photo-1505740420928-5e560c06d30e?w=500"
          },
          "category": {
            "type": "string",
            "example": "Electronics"
          },
          "price": {
            "type": "number",
            "example": 89.99
          },
          "sizes": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "example": ["S", "M", "L"]
          },
          "ratings": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Rating"
            }
          }
        }
      },
      "Rating": {
        "type": "object",
        "properties": {
          "userID": {
            "type": "string",
            "example": "2"
          },
          "rating": {
            "type": "number",
            "minimum": 1,
            "maximum": 5,
            "example": 4
          }
        }
      },
      "CartItem": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "example": 1
          },
          "productID": {
            "type": "integer",
            "example": 1
          },
          "userID": {
            "type": "integer",
            "example": 2
          },
          "quantity": {
            "type": "integer",
            "example": 1
          }
        }
      },
...
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "message": {
            "type": "string"
          }
        }
      },
...
    }
  },
  "paths": {
...
    "/api/products": {
      "get": {
        "tags": ["Products"],
        "summary": "Get All Products",
        "description": "Retrieve all products",
        "security": [
          {
            "JWT": []
          }
        ],
        "responses": {
          "200": {
            "description": "List of products",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Product"
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": ["Products"],
        "summary": "Add Product",
        "description": "Add a new product with image upload",
        "security": [
          {
            "JWT": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "required": ["name", "price", "imageUrl"],
                "properties": {
                  "name": {
                    "type": "string",
                    "description": "Product name"
                  },
                  "price": {
                    "type": "number",
                    "description": "Product price"
                  },
                  "sizes": {
                    "type": "string",
                    "description": "Product sizes (comma-separated)"
                  },
                  "imageUrl": {
                    "type": "string",
                    "format": "binary",
                    "description": "Product image file"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Product created successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Product"
                }
              }
            }
          },
          "400": {
            "description": "Bad request - missing required fields",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/cart": {
      "get": {
        "tags": ["Cart"],
        "summary": "Get Cart Items",
        "description": "Retrieve all cart items",
        "security": [
          {
            "JWT": []
          }
        ],
        "responses": {
          "200": {
            "description": "List of cart items",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/CartItem"
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": ["Cart"],
        "summary": "Add to Cart",
        "description": "Add an item to the cart",
        "security": [
          {
            "JWT": []
          }
        ],
        "parameters": [
          {
            "in": "query",
            "name": "productID",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Product ID to add to cart"
          },
          {
            "in": "query",
            "name": "quantity",
            "required": true,
            "schema": {
              "type": "integer"
            },
            "description": "Quantity of the product"
          }
        ],
        "responses": {
          "201": {
            "description": "Item added to cart successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CartItem"
                }
              }
            }
          },
          "400": {
            "description": "Bad request - missing required parameters",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
  }
}